revert modal stuff
