+++ src/lib/components/navigation/NavMenu/LightSwitch.svelte
@@ -65,3 +65,18 @@
         {/each}
     </SegmentedControl.Control>
 </SegmentedControl>
+
+<svelte:head>
+    <script>
+        (function initMode() {
+            let mode =
+                typeof localStorage !== "undefined" ? JSON.parse(localStorage.getItem("mode")) || "system" : "system";
+            if (mode === "system") {
+                const pref = window.matchMedia("(prefers-color-scheme: dark)").matches ? "dark" : "light";
+                document.documentElement.setAttribute("data-mode", pref);
+            } else {
+                document.documentElement.setAttribute("data-mode", mode);
+            }
+        })();
+    </script>
+</svelte:head>
