+++ src/lib/components/wishlists/ItemDrawer.svelte
@@ -6,7 +6,7 @@
     import { getFormatter } from "$lib/i18n";
     import ItemImage from "./ItemCard/components/ItemImage.svelte";
     import ItemAttributes from "./ItemCard/components/ItemAttributes.svelte";
-    import { Dialog, Portal, useDialog } from "@skeletonlabs/skeleton-svelte";
+    import { Dialog, Portal } from "@skeletonlabs/skeleton-svelte";
     import ModalContent from "../modals/parts/ModalContent.svelte";
     import ModalBackdrop from "../modals/parts/ModalBackdrop.svelte";
     import type { InternalItemCardProps } from "./ItemCard/ItemCard.svelte";
@@ -14,28 +14,17 @@
     let { defaultImage: _defaultImage, item, ...props }: InternalItemCardProps = $props();
     const t = getFormatter();
 
-    const id = $props.id();
-    const dialog = useDialog({ id });
-
-    let mount = $state(false);
+    let open = $state(false);
 
     $effect(() => {
         if (page.url.searchParams.get("item-id") === item.id.toString()) {
-            dialog().setOpen(true);
-        }
-    });
-
-    $effect(() => {
-        if (dialog().open) {
-            mount = true;
-        } else {
-            setTimeout(() => (mount = false), 250);
+            open = true;
         }
     });
 
     async function closeDialog() {
         await goto(page.url.pathname, { replaceState: true, noScroll: true });
-        dialog().setOpen(false);
+        open = false;
     }
 
     const transitionIn =
@@ -44,49 +33,47 @@
         "data-[state=closed]:translate-y-1/2 starting:data-[state=closed]:translate-y-0 md:data-[state=closed]:translate-y-0 md:data-[state=closed]:scale-90 md:starting:data-[state=closed]:scale-100";
 </script>
 
-<Dialog.Provider value={dialog}>
-    {#if mount}
-        <Portal>
-            <ModalBackdrop class="duration-250"></ModalBackdrop>
-            <Dialog.Positioner class="fixed inset-0 z-50 flex items-end justify-center sm:mx-4 sm:items-center">
-                <ModalContent
-                    class="md:rounded-b-container max-h-3/4 translate-x-0 rounded-b-none duration-250 {transitionIn} {transitionOut}"
-                >
-                    <div class="grid grid-cols-[1fr_auto] justify-between gap-2 pb-2">
-                        <Dialog.Title class="truncate text-xl font-bold text-wrap wrap-break-word md:text-2xl">
-                            {item.name}
-                        </Dialog.Title>
-                        <Dialog.CloseTrigger
-                            class="preset-tonal-surface border-surface-500 btn-icon border"
-                            aria-label={$t("a11y.close")}
-                            onclick={closeDialog}
-                        >
-                            <iconify-icon icon="ion:close"></iconify-icon>
-                        </Dialog.CloseTrigger>
-                    </div>
+<Dialog onOpenChange={(e) => (open = e.open)} {open}>
+    <Portal>
+        <ModalBackdrop class="duration-250"></ModalBackdrop>
+        <Dialog.Positioner class="fixed inset-0 z-50 flex items-end justify-center sm:mx-4 sm:items-center">
+            <ModalContent
+                class="md:rounded-b-container max-h-3/4 translate-x-0 rounded-b-none duration-250 {transitionIn} {transitionOut}"
+            >
+                <div class="grid grid-cols-[1fr_auto] justify-between gap-2 pb-2">
+                    <Dialog.Title class="truncate text-xl font-bold text-wrap wrap-break-word md:text-2xl">
+                        {item.name}
+                    </Dialog.Title>
+                    <Dialog.CloseTrigger
+                        class="preset-tonal-surface border-surface-500 btn-icon border"
+                        aria-label={$t("a11y.close")}
+                        onclick={closeDialog}
+                    >
+                        <iconify-icon icon="ion:close"></iconify-icon>
+                    </Dialog.CloseTrigger>
+                </div>
 
-                    <div class="flex max-h-[40dvh] justify-center">
-                        <ItemImage class="max-h-full object-scale-down" {item}>
-                            {#snippet defaultImage(t)}
-                                {@render _defaultImage(t, "w-40 md:w-48 aspect-square rounded")}
-                            {/snippet}
-                        </ItemImage>
-                    </div>
+                <div class="flex max-h-[40dvh] justify-center">
+                    <ItemImage class="max-h-full object-scale-down" {item}>
+                        {#snippet defaultImage(t)}
+                            {@render _defaultImage(t, "w-40 md:w-48 aspect-square rounded")}
+                        {/snippet}
+                    </ItemImage>
+                </div>
 
-                    {#if item.url}
-                        <a class="anchor dark:text-primary-200!" href={item.url} rel="noreferrer" target="_blank">
-                            {$t("wishes.view-item")}
-                        </a>
-                    {/if}
+                {#if item.url}
+                    <a class="anchor dark:text-primary-200!" href={item.url} rel="noreferrer" target="_blank">
+                        {$t("wishes.view-item")}
+                    </a>
+                {/if}
 
-                    <ItemAttributes {...props} fullNotes {item} showDetail />
+                <ItemAttributes {...props} fullNotes {item} showDetail />
 
-                    <div class="flex flex-row justify-between">
-                        <ClaimButtons {item} {...props} />
-                        <ManageButtons {item} {...props} onEdit={() => dialog().setOpen(false)} />
-                    </div>
-                </ModalContent>
-            </Dialog.Positioner>
-        </Portal>
-    {/if}
-</Dialog.Provider>
+                <div class="flex flex-row justify-between">
+                    <ClaimButtons {item} {...props} />
+                    <ManageButtons {item} {...props} onEdit={() => (open = false)} />
+                </div>
+            </ModalContent>
+        </Dialog.Positioner>
+    </Portal>
+</Dialog>
